+++ src/JsonApi/Schema/Error.php
@@ -49,14 +49,14 @@ class Error
      */
     public static function createFromArray(array $error)
     {
-        $id = empty($error["id"]) === false ? $error["id"] : "";
+        $id = isset($error["id"]) && is_string($error["id"]) ? $error["id"] : "";
         $meta = isset($error["meta"]) && is_array($error["meta"]) ? $error["meta"] : [];
         $links = Links::createFromArray(isset($error["links"]) && is_array($error["links"]) ? $error["links"] : []);
-        $status = empty($error["status"]) === false ? $error["status"] : "";
-        $code = empty($error["code"]) === false ? $error["code"] : "";
-        $title = empty($error["title"]) === false ? $error["title"] : "";
-        $detail = empty($error["detail"]) === false ? $error["detail"] : "";
-        $source = ErrorSource::fromArray(isset($error["source"]) && is_array($error["source"]) ? $error["source"] : []);
+        $status = isset($error["status"]) && is_scalar($error["status"]) ? (string) $error["status"] : "";
+        $code = isset($error["code"]) && is_string($error["code"]) ? $error["code"] : "";
+        $title = isset($error["title"]) && is_string($error["title"]) ? $error["title"] : "";
+        $detail = isset($error["detail"]) && is_string($error["detail"]) ? $error["detail"] : "";
+        $source = ErrorSource::createFromArray(isset($error["source"]) && is_array($error["source"]) ? $error["source"] : []);
 
         return new self($id, $meta, $links, $status, $code, $title, $detail, $source);
     }